refactor: 서비스 분리 및 네이밍 변경
