+++ src/main/java/racingcar/controller/ApiController.java
@@ -9,7 +9,8 @@ import org.springframework.web.bind.annotation.RestController;
 import racingcar.dto.request.GameRequestDto;
 import racingcar.dto.response.GameResponseDto;
 import racingcar.dto.response.GameResultDto;
-import racingcar.service.GameService;
+import racingcar.service.GameFindService;
+import racingcar.service.GamePlayService;
 
 import javax.validation.Valid;
 import java.net.URI;
@@ -18,21 +19,23 @@ import java.util.List;
 @RestController
 public class ApiController {
 
-    private final GameService gameService;
+    private final GameFindService gameFindService;
+    private final GamePlayService gamePlayService;
 
-    public ApiController(final GameService gameService) {
-        this.gameService = gameService;
+    public ApiController(final GameFindService gameService, GamePlayService gamePlayService) {
+        this.gameFindService = gameService;
+        this.gamePlayService = gamePlayService;
     }
 
     @PostMapping(value = "/plays", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<GameResponseDto> playGame(@Valid @RequestBody GameRequestDto gameRequestDto) {
-        final GameResponseDto gameResponseDto = gameService.playGame(gameRequestDto);
+        final GameResponseDto gameResponseDto = gamePlayService.playGame(gameRequestDto);
         return ResponseEntity.created(URI.create("/plays")).body(gameResponseDto);
     }
 
     @GetMapping("/plays")
     public List<GameResultDto> retrieveGameResults() {
-        List<GameResultDto> gameResultDtos = gameService.findAllGames();
+        List<GameResultDto> gameResultDtos = gameFindService.findAllGames();
         return gameResultDtos;
     }
 }
